2. Передача информации о заказе на сервер Dish через брокер Kafka

+++ src/main/java/ru/job4j/job4j_order/repository/OrderRepository.java
@@ -3,5 +3,5 @@ package ru.job4j.job4j_order.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.job4j.job4j_order.model.Order;
 
-public interface OrderRepository extends CrudRepository<Order, Integer> {
+public interface OrderRepository extends CrudRepository<Order, Long> {
 }
